refactor: 사용자 임계값 수정 검증 보완

+++ backend-spring/src/main/java/com/example/factoryguard/application/dto/user/UpdateThresholdCommand.java
@@ -3,12 +3,14 @@ package com.example.factoryguard.application.dto.user;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Getter
 @RequiredArgsConstructor
 public class UpdateThresholdCommand {
 
-    private final double anomalyThreshold;
-    private final double lowConfidenceThreshold;
+    private final BigDecimal anomalyThreshold;
+    private final BigDecimal lowConfidenceThreshold;
     private final String applyScope;
     private final String changeReason;
 }
